requirements.txt
# .\venv\Scripts\Activate.ps1
# Lists all python dependencies for pip install -r requirements.txt

# git add .
# git commit -m "your message"
# git push

fastapi
pydantic
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
pandas